dynamic routing

+++ utils.py
@@ -196,7 +196,7 @@ def get_region_boxes_v2(output, n_models, conf_thresh, num_classes, anchors, num
     cs = n_models
     nA = num_anchors
     nC = num_classes
-    anchor_step = len(anchors)/num_anchors
+    anchor_step = int(len(anchors)/num_anchors)
     if output.dim() == 3:
         output = output.unsqueeze(0)
     batch = output.size(0)
